+++ bmc-objectstorage/src/main/java/com/oracle/bmc/objectstorage/model/CreateBucketDetails.java
@@ -37,13 +37,19 @@ public class CreateBucketDetails {
         @JsonProperty("metadata")
         private Map<String, String> metadata;
 
+        @JsonProperty("publicAccessType")
+        private PublicAccessType publicAccessType;
+
         public CreateBucketDetails build() {
-            return new CreateBucketDetails(name, compartmentId, metadata);
+            return new CreateBucketDetails(name, compartmentId, metadata, publicAccessType);
         }
 
         @JsonIgnore
         public Builder copy(CreateBucketDetails o) {
-            return name(o.getName()).compartmentId(o.getCompartmentId()).metadata(o.getMetadata());
+            return name(o.getName())
+                    .compartmentId(o.getCompartmentId())
+                    .metadata(o.getMetadata())
+                    .publicAccessType(o.getPublicAccessType());
         }
     }
 
@@ -78,4 +84,52 @@ public class CreateBucketDetails {
      **/
     @JsonProperty("metadata")
     Map<String, String> metadata;
+    /**
+     * The type of public access available on this bucket. Allows authenticated caller to access the bucket or
+     * contents of this bucket. By default a bucket is set to NoPublicAccess. It is treated as NoPublicAccess
+     * when this value is not specified. When the type is NoPublicAccess the bucket does not allow any public access.
+     * When the type is ObjectRead the bucket allows public access to the GetObject, HeadObject, ListObjects.
+     *
+     **/
+    public enum PublicAccessType {
+        NoPublicAccess("NoPublicAccess"),
+        ObjectRead("ObjectRead"),
+        ;
+
+        private final String value;
+        private static Map<String, PublicAccessType> map;
+
+        static {
+            map = new HashMap<>();
+            for (PublicAccessType v : PublicAccessType.values()) {
+                map.put(v.getValue(), v);
+            }
+        }
+
+        PublicAccessType(String value) {
+            this.value = value;
+        }
+
+        @JsonValue
+        public String getValue() {
+            return value;
+        }
+
+        @JsonCreator
+        public static PublicAccessType create(String key) {
+            if (map.containsKey(key)) {
+                return map.get(key);
+            }
+            throw new RuntimeException("Invalid PublicAccessType: " + key);
+        }
+    };
+    /**
+     * The type of public access available on this bucket. Allows authenticated caller to access the bucket or
+     * contents of this bucket. By default a bucket is set to NoPublicAccess. It is treated as NoPublicAccess
+     * when this value is not specified. When the type is NoPublicAccess the bucket does not allow any public access.
+     * When the type is ObjectRead the bucket allows public access to the GetObject, HeadObject, ListObjects.
+     *
+     **/
+    @JsonProperty("publicAccessType")
+    PublicAccessType publicAccessType;
 }